@@ -1,4 +1,4 @@
-export * from './User';
-export * from './Lock';
-export * from './Paym';
-export * from './Invo';
+export * from './User.js';
+export * from './Lock.js';
+export * from './Paym.js';
+export * from './Invo.js';
