@@ -1,15 +1,19 @@
+import express from 'express';
+import helmet from 'helmet';
+import morgan from 'morgan';
+import { v4 as uuidv4 } from 'uuid';
+import logger from './utils/logger.js';
+import config from './config.js';
+import rateLimit from 'express-rate-limit';
+import apiController from './controllers/api.js';
+import siteController from './controllers/website.js';
+
 process.on('uncaughtException', function (err) {
   console.error(err);
   console.log('Node NOT Exiting...');
 });
 
 process.env.NODE_TLS_REJECT_UNAUTHORIZED = '0';
-let express = require('express');
-const helmet = require('helmet');
-let morgan = require('morgan');
-import { v4 as uuidv4 } from 'uuid';
-let logger = require('./utils/logger');
-const config = require('./config');
 
 morgan.token('id', function getId(req) {
   return req.id;
@@ -20,7 +24,6 @@ app.enable('trust proxy');
 app.use(helmet.hsts());
 app.use(helmet.hidePoweredBy());
 
-const rateLimit = require('express-rate-limit');
 const limiter = rateLimit({
   windowMs: 15 * 60 * 1000,
   max: config.rateLimit || 200,
@@ -38,14 +41,13 @@ app.use(
   ),
 );
 
-let bodyParser = require('body-parser');
 
-app.use(bodyParser.urlencoded({ extended: false })); // parse application/x-www-form-urlencoded
-app.use(bodyParser.json(null)); // parse application/json
+app.use(express.urlencoded({ extended: false })); // parse application/x-www-form-urlencoded
+app.use(express.json(null)); // parse application/json
 
 app.use('/static', express.static('static'));
-app.use(require('./controllers/api'));
-app.use(require('./controllers/website'));
+app.use(apiController);
+app.use(siteController);
 
 const bindHost = process.env.HOST || '0.0.0.0';
 const bindPort = process.env.PORT || 3000;
